Change timing points' start time type to `double`
Encountered a case where this is a `double`: /b/1493524.

+++ src/com/rian/difficultycalculator/beatmap/timings/ControlPoint.java
@@ -7,12 +7,12 @@ public abstract class ControlPoint {
     /**
      * The time at which this control point takes effect, in milliseconds.
      */
-    public final int time;
+    public final double time;
 
     /**
      * @param time The time at which this control point takes effect, in milliseconds.
      */
-    public ControlPoint(int time) {
+    public ControlPoint(double time) {
         this.time = time;
     }
